+++ src/mocks/validator/MockRoninValidatorSetExtended.sol
@@ -21,7 +21,9 @@ contract MockRoninValidatorSetExtended is MockRoninValidatorSetOverridePrecompil
     _epochs.push(block.number);
   }
 
-  function epochOf(uint256 _block) public view override returns (uint256 _epoch) {
+  function epochOf(
+    uint256 _block
+  ) public view override returns (uint256 _epoch) {
     for (uint256 _i = _epochs.length; _i > 0; _i--) {
       if (_block > _epochs[_i - 1]) {
         return _i;
@@ -29,8 +31,10 @@ contract MockRoninValidatorSetExtended is MockRoninValidatorSetOverridePrecompil
     }
   }
 
-  function epochEndingAt(uint256 _block) public view override(ITimingInfo, TimingStorage) returns (bool) {
-    for (uint _i = 0; _i < _epochs.length; _i++) {
+  function epochEndingAt(
+    uint256 _block
+  ) public view override(ITimingInfo, TimingStorage) returns (bool) {
+    for (uint256 _i = 0; _i < _epochs.length; _i++) {
       if (_block == _epochs[_i]) {
         return true;
       }
@@ -38,15 +42,19 @@ contract MockRoninValidatorSetExtended is MockRoninValidatorSetOverridePrecompil
     return false;
   }
 
-  function getJailUntils(address[] calldata _addrs) public view returns (uint256[] memory jailUntils_) {
+  function getJailUntils(
+    address[] calldata _addrs
+  ) public view returns (uint256[] memory jailUntils_) {
     jailUntils_ = new uint256[](_addrs.length);
-    for (uint _i = 0; _i < _addrs.length; _i++) {
+    for (uint256 _i = 0; _i < _addrs.length; _i++) {
       jailUntils_[_i] = _blockProducerJailedBlock[_addrs[_i]];
     }
   }
 
-  function addValidators(address[] calldata _addrs) public {
-    for (uint _i = 0; _i < _addrs.length; _i++) {
+  function addValidators(
+    address[] calldata _addrs
+  ) public {
+    for (uint256 _i = 0; _i < _addrs.length; _i++) {
       address _cAddr = _addrs[_i];
       _validatorIds[_i] = _cAddr;
       _validatorMap[_cAddr] = true;
