+++ src/ronin/slash-indicator/CreditScore.sol
@@ -2,14 +2,16 @@
 
 pragma solidity ^0.8.9;
 
-import "../../interfaces/IMaintenance.sol";
-import "../../interfaces/validator/IRoninValidatorSet.sol";
-import "../../interfaces/slash-indicator/ICreditScore.sol";
 import "../../extensions/collections/HasContracts.sol";
 import "../../extensions/consumers/PercentageConsumer.sol";
+import "../../interfaces/IMaintenance.sol";
+import "../../interfaces/slash-indicator/ICreditScore.sol";
+import "../../interfaces/validator/IRoninValidatorSet.sol";
+
 import "../../libraries/Math.sol";
-import { HasValidatorDeprecated, HasMaintenanceDeprecated } from "../../utils/DeprecatedSlots.sol";
+
 import { ErrUnauthorized, RoleAccess } from "../../utils/CommonErrors.sol";
+import { HasMaintenanceDeprecated, HasValidatorDeprecated } from "../../utils/DeprecatedSlots.sol";
 
 abstract contract CreditScore is
   ICreditScore,
@@ -54,7 +56,7 @@ abstract contract CreditScore is
     );
     uint256[] memory updatedCreditScores = new uint256[](validatorIds.length);
 
-    for (uint i = 0; i < validatorIds.length;) {
+    for (uint256 i = 0; i < validatorIds.length;) {
       address vId = validatorIds[i];
 
       uint256 indicator = _getUnavailabilityIndicatorById(vId, period);
@@ -74,13 +76,11 @@ abstract contract CreditScore is
     emit CreditScoresUpdated(validatorIds, updatedCreditScores);
   }
 
-  function execResetCreditScores(address[] calldata validatorIds)
-    external
-    override
-    onlyContract(ContractType.VALIDATOR)
-  {
+  function execResetCreditScores(
+    address[] calldata validatorIds
+  ) external override onlyContract(ContractType.VALIDATOR) {
     uint256[] memory updatedCreditScores = new uint256[](validatorIds.length);
-    for (uint i = 0; i < validatorIds.length;) {
+    for (uint256 i = 0; i < validatorIds.length;) {
       address _validator = validatorIds[i];
       delete _creditScore[_validator];
       delete updatedCreditScores[i];
@@ -95,7 +95,9 @@ abstract contract CreditScore is
   /**
    * @inheritdoc ICreditScore
    */
-  function bailOut(TConsensus consensusAddr) external override {
+  function bailOut(
+    TConsensus consensusAddr
+  ) external override {
     address validatorId = __css2cid(consensusAddr);
     IRoninValidatorSet validatorContract = IRoninValidatorSet(getContract(ContractType.VALIDATOR));
     if (!validatorContract.isValidatorCandidate(consensusAddr)) {
@@ -161,23 +163,22 @@ abstract contract CreditScore is
   /**
    * @inheritdoc ICreditScore
    */
-  function getCreditScore(TConsensus consensusAddr) external view override returns (uint256) {
+  function getCreditScore(
+    TConsensus consensusAddr
+  ) external view override returns (uint256) {
     return _creditScore[__css2cid(consensusAddr)];
   }
 
   /**
    * @inheritdoc ICreditScore
    */
-  function getManyCreditScores(TConsensus[] calldata consensusAddrs)
-    public
-    view
-    override
-    returns (uint256[] memory resultList)
-  {
+  function getManyCreditScores(
+    TConsensus[] calldata consensusAddrs
+  ) public view override returns (uint256[] memory resultList) {
     address[] memory validatorIds = __css2cidBatch(consensusAddrs);
     resultList = new uint256[](validatorIds.length);
 
-    for (uint i = 0; i < resultList.length;) {
+    for (uint256 i = 0; i < resultList.length;) {
       resultList[i] = _creditScore[validatorIds[i]];
 
       unchecked {
@@ -202,9 +203,13 @@ abstract contract CreditScore is
    */
   function _setUnavailabilityIndicator(address _validator, uint256 period, uint256 _indicator) internal virtual;
 
-  function __css2cid(TConsensus consensusAddr) internal view virtual returns (address);
+  function __css2cid(
+    TConsensus consensusAddr
+  ) internal view virtual returns (address);
 
-  function __css2cidBatch(TConsensus[] memory consensusAddrs) internal view virtual returns (address[] memory);
+  function __css2cidBatch(
+    TConsensus[] memory consensusAddrs
+  ) internal view virtual returns (address[] memory);
 
   /**
    * @dev See `ICreditScore-setCreditScoreConfigs`.
